+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.1.3 / 2023-02-12
   * Added `grin`, `horns`, `open` and `sad` emojis.
+  * Now based on [`JekyllSupport::JekyllTag`](https://mslinn.com/jekyll/10200-jekyll-plugin-support.html) instead of `Liquid::Tag`
+
 
 ## 0.1.3 / 2022-04-15
   * Renamed `lib/jekyll_tag_plugins.rb` to  `lib/jekyll_block_tag_plugin.rb`